# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.IngestSleepSegmentsRequest:
    properties:
      segments:
        items:
          $ref: '#/definitions/handlers.SleepSegment'
        type: array
    type: object
  handlers.IngestSteps15mRequest:
    properties:
      points:
        items:
          $ref: '#/definitions/handlers.Step15mPoint'
        type: array
    type: object
...
  handlers.SleepSegment:
    properties:
      end_utc:
        description: RFC3339
        type: string
      sleep_date:
        description: YYYY-MM-DD
        type: string
      source:
        type: string
      stage:
        description: deep/light/rem/awake/off_wrist/unknown
        type: string
      start_utc:
        description: RFC3339
        type: string
      tz_offset_min:
        type: integer
    type: object
  handlers.SleepSummaryRequest:
    properties:
      awake_s:
        type: integer
      deep_s:
        type: integer
      light_s:
        type: integer
      rem_s:
        type: integer
      sleep_date:
        description: YYYY-MM-DD (device-local night)
        type: string
      sleep_start_utc:
        description: RFC3339 optional
        type: string
      source:
        type: string
      total_sleep_s:
        type: integer
      tz_offset_min:
        type: integer
      wake_utc:
        description: RFC3339 optional
        type: string
    type: object
...
  handlers.Step15mPoint:
    properties:
      calories:
        type: integer
      device_date:
        description: YYYY-MM-DD (device-local)
        type: string
      distance_m:
        type: integer
      run_steps:
        type: integer
      source:
        description: device/derived/manual/import (optional)
        type: string
      sport_duration_s:
        description: active duration in seconds
        type: integer
      time_index:
        description: 0..95
        type: integer
      ts_utc:
        description: RFC3339
        type: string
      tz_offset_min:
        type: integer
      walk_steps:
        type: integer
    type: object
...
  handlers.UpsertResponse:
    properties:
      upserted:
        example: 1
        type: integer
    type: object
...
  httpx.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  store.SleepSegmentResponse:
    properties:
      end_utc:
        type: string
      sleep_date:
        type: string
      source:
        type: string
      stage:
        type: string
      start_utc:
        type: string
      synced_at:
        type: string
      tz_offset_min:
        type: integer
    type: object
  store.SleepSummaryResponse:
    properties:
      awake_s:
        type: integer
      deep_s:
        type: integer
      device_id:
        type: string
      light_s:
        type: integer
      rem_s:
        type: integer
      score:
        type: integer
      sleep_date:
        type: string
      sleep_start_utc:
        type: string
      source:
        type: string
      stage_data:
        items:
          type: integer
        type: array
      synced_at:
        type: string
      total_sleep_s:
        type: integer
      tz_offset_min:
        type: integer
      wake_utc:
        type: string
    type: object
  store.Steps15mResponse:
    properties:
      calories:
        type: integer
      device_date:
        type: string
      distance_m:
        type: integer
      run_steps:
        type: integer
      sport_duration_s:
        type: integer
      synced_at:
        type: string
      time_index:
        type: integer
      ts_utc:
        type: string
      tz_offset_min:
        type: integer
      walk_steps:
        type: integer
    type: object
  store.StepsDailyRow:
    properties:
      calories:
        type: integer
      day_utc:
        type: string
      distance_m:
        type: integer
      last_synced_at:
        type: string
      running_steps:
        type: integer
      total_steps:
        type: integer
    type: object
...
info:
  contact: {}
  description: Backend API for AGEM wearable users, devices, qring sensor data, sleep,
    and workouts.
  title: AGEM Backend API
  version: "1.0"
paths:
...
  /v1/devices/{device_id}/sleep/segments:
    get:
      description: Lists sleep stage segments by device-local sleep date.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Device-local sleep date
        format: date
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.SleepSegmentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List sleep segments
      tags:
      - Sleep
    post:
      consumes:
      - application/json
      description: Idempotently ingests sleep stage segments for a device.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Sleep segment payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IngestSleepSegmentsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UpsertResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Ingest sleep segments
      tags:
      - Sleep
  /v1/devices/{device_id}/sleep/summary:
    get:
      description: Returns a sleep summary by device-local sleep date.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Device-local sleep date
        format: date
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.SleepSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Get sleep summary
      tags:
      - Sleep
    post:
      consumes:
      - application/json
      description: Creates or updates one device-local sleep summary for a device
        and sleep date.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Sleep summary payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SleepSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.SleepSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Upsert sleep summary
      tags:
      - Sleep
  /v1/devices/{device_id}/steps/daily:
    get:
      description: Aggregates daily step totals from the steps_15m time series collection.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Inclusive RFC3339 start timestamp
        format: date-time
        in: query
        name: from
        required: true
        type: string
      - description: Exclusive RFC3339 end timestamp
        format: date-time
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.StepsDailyRow'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Aggregate daily steps
      tags:
      - Steps
  /v1/devices/{device_id}/steps15m:
    get:
      description: Lists 15-minute step points by device-local date.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Device-local date
        format: date
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Steps15mResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List 15-minute steps
      tags:
      - Steps
    post:
      consumes:
      - application/json
      description: Idempotently ingests 15-minute qring step detail points.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Step points
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IngestSteps15mRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UpsertResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Ingest 15-minute steps
      tags:
      - Steps
...
swagger: "2.0"